+++ packages/react-querybuilder/src/components/Rule.tsx
@@ -208,6 +208,7 @@ export const RuleComponents: React.MemoExoticComponent<
           title={r.translations.dragHandle.title}
           label={r.translations.dragHandle.label}
           className={r.classNames.dragHandle}
+          dragHandleProps={r.dragHandleProps}
           ruleOrGroup={r.rule}
         />
       )}